+++ src-tauri/src/platforms/bilibili/legacy.rs
@@ -211,6 +211,7 @@ pub async fn download(
         opts.download_subtitles,
         &extra,
         opts.audio_format.as_deref(),
+        opts.save_encrypted_hls,
     )
     .await
 }
@@ -266,6 +267,7 @@ async fn download_playlist(
             opts.download_subtitles,
             &extra,
             opts.audio_format.as_deref(),
+            opts.save_encrypted_hls,
         )
         .await
         {